# app-service/backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/v1/images/{id}/media:
    get:
      tags:
      - image-rest-controller
      operationId: getImageFile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - image-rest-controller
      operationId: replaceImageFile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/solutions/{id}/images:
    post:
      tags:
      - solution-rest-controller
      operationId: uploadSolutionImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/solutions/{id}/comment:
    post:
      tags:
      - solution-rest-controller
      operationId: postComment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentPostDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/solutions/{id}:
    get:
      tags:
      - solution-rest-controller
      operationId: getSolutionById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SolutionDTO"
    delete:
      tags:
      - solution-rest-controller
      operationId: deleteSolutionById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/solutions/{id}/pdfs/:
    get:
      tags:
      - solution-rest-controller
      operationId: createSolutionPDF
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
  /api/v1/solutions/{id}/comments:
    get:
      tags:
      - solution-rest-controller
      operationId: getCommentsForSolution
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentDTO"
...
components:
  schemas:
    ImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
...
    CommentPostDTO:
      type: object
      properties:
        text:
          type: string
...
    UserBasicInfoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
...
    CommentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
        lastUpdate:
          type: string
          format: date-time
        owner:
          $ref: "#/components/schemas/UserBasicInfoDTO"
        solution:
          $ref: "#/components/schemas/SolutionBasicInfoDTO"
    ExerciseBasicInfoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        numSolutions:
          type: integer
          format: int32
    SolutionBasicInfoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    SolutionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        owner:
          $ref: "#/components/schemas/UserBasicInfoDTO"
        lastUpdate:
          type: string
          format: date-time
        description:
          type: string
        exercise:
          $ref: "#/components/schemas/ExerciseBasicInfoDTO"
        solImage:
          $ref: "#/components/schemas/ImageDTO"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/CommentDTO"
...